Delete unnecessary test. Re-organise code.
