--- src/test/kotlin/com/hamersztein/chorserver/TestChorServerApplication.kt
@@ -1,9 +1,0 @@
-package com.hamersztein.chorserver
-
-import org.springframework.boot.fromApplication
-import org.springframework.boot.with
-
-
-fun main(args: Array<String>) {
-	fromApplication<ChorServerApplication>().with(TestcontainersConfiguration::class).run(*args)
-}
