+++ src/test/kotlin/com/hamersztein/chorserver/infrastructure/RepositoryTest.kt
@@ -1,6 +1,5 @@
 package com.hamersztein.chorserver.infrastructure
 
-import com.hamersztein.chorserver.TestcontainersConfiguration
 import com.hamersztein.chorserver.shared.GeometryFactoryConfiguration
 import com.hamersztein.chorserver.shared.R2DBCMappingConfig
 import org.springframework.boot.test.autoconfigure.data.r2dbc.DataR2dbcTest
