+++ src/test/kotlin/com/hamersztein/chorserver/infrastructure/TestcontainersConfiguration.kt
@@ -1,4 +1,4 @@
-package com.hamersztein.chorserver
+package com.hamersztein.chorserver.infrastructure
 
 import org.springframework.boot.test.context.TestConfiguration
 import org.springframework.boot.testcontainers.service.connection.ServiceConnection
